Updated BusTimetableModel to use LONG for Id and BusLineId properties instead of INT

+++ app/src/main/java/com/example/ratbvkotlin/App.kt
@@ -102,7 +102,7 @@ class App : Application() {
         }
         viewModel { (
                     scheduleLink: String,
-                    busStationId: Int,
+                    busStationId: Long,
                     busStationName: String
                 ) ->
             BusTimetablesViewModel(
